+++ familyhelper-ui-view/familyhelper-ui-view-pc/src/views/items/systemSettings/permissionNode/PermissionNode.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="permission-node-container">
-    <content-panel
-      class="content-panel"
+    <border-layout-panel
+      class="border-layout-panel"
       :header-visible="true"
     >
       <table-panel
@@ -45,7 +45,7 @@
           新建权限
         </el-button>
       </div>
-    </content-panel>
+    </border-layout-panel>
     <entity-maintain-dialog
       :mode="dialogMode"
       :visible.sync="dialogVisible"
@@ -86,7 +86,7 @@
 </template>
 
 <script>
-import ContentPanel from '@/components/layout/LayoutPanel.vue';
+import BorderLayoutPanel from '@/components/layout/BorderLayoutPanel.vue';
 import TablePanel from '@/components/layout/TablePanel.vue';
 import EntityMaintainDialog from '@/components/dialog/EntityMaintainDialog.vue';
 
@@ -98,7 +98,7 @@ import resolveResponse from '@/util/response';
 
 export default {
   name: 'PermissionNode',
-  components: { EntityMaintainDialog, ContentPanel, TablePanel },
+  components: { EntityMaintainDialog, BorderLayoutPanel, TablePanel },
   data() {
     const keyValidator = (rule, value, callback) => {
       Promise.resolve(value)